requirements.txt
pip==20.2.4
gevent --pre
auto-py-to-exe
arrow==0.17.0
db-dtypes==1.0.4
duckdb==0.8.0
fsspec==0.8.4
gcsfs==0.7.1
google-cloud-bigquery==3.4.0
google-cloud-bigquery-storage==2.13.2
google-cloud-storage==1.30.0
grpcio==1.47.0
grpcio-gcp==0.2.2
Jinja2==2.11.2
kfp==1.8.14
MarkupSafe==1.1.1
numpy==1.18.5
numba==0.56.4
openpyxl==3.0.7
pandas==1.1.4
PuLP==2.4
pyarrow==8.0.0
scikit-learn==0.23.2
statsmodels==0.12.1
XlsxWriter==1.3.7
xlrd==1.2.0
gurobipy==10.0.3
urllib3==1.26.15
scipy==1.7.3
python-dateutil==2.9.0
matplotlib==3.3.1
